pyproject.toml
[project]
name = "flight-chatbot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.9",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "anyio==4.6.0",
    "attrs==24.2.0",
    "bitsandbytes==0.43.1",
    "cachetools==5.5.2",
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "filelock==3.16.1",
    "filetype==1.2.0",
    "frozenlist==1.4.1",
    "fsspec==2024.9.0",
    "google-ai-generativelanguage==0.6.17",
    "google-api-core==2.24.2",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.69.2",
    "greenlet==3.1.1",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.14.0",
    "httpcore==1.0.6",
    "httpx==0.27.2",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.25.1",
    "idna==3.10",
    "intel-openmp==2021.4.0",
    "jinja2==3.1.4",
    "jiter==0.7.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.7",
    "langchain-community==0.3.7",
    "langchain-core==0.3.48",
    "langchain-google-genai==2.1.1",
    "langchain-ollama==0.2.0",
    "langchain-openai==0.2.9",
    "langchain-text-splitters==0.3.2",
    "langsmith==0.1.132",
    "markupsafe==3.0.0",
    "marshmallow==3.22.0",
    "mkl==2021.4.0",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "networkx==3.3",
    "numpy==1.26.4",
    "ollama==0.4.0",
    "openai==1.55.0",
    "orjson==3.10.7",
    "packaging==24.1",
    "propcache==0.2.0",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pydantic-settings==2.6.1",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "regex==2024.9.11",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rsa==4.9",
    "safetensors==0.4.5",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.35",
    "sympy==1.13.3",
    "tbb==2021.13.1",
    "tenacity==8.5.0",
    "tiktoken==0.8.0",
    "tokenizers==0.19.1",
    "torch==2.3.1",
    "tqdm==4.66.5",
    "transformers==4.42.4",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "urllib3==2.2.3",
    "yarl==1.14.0",
]
